integration_tests/requirements.txt
python-socketio>=5
websocket-client
pytest